ui: remove dependency from Timeline and State
Today there is a dependency between Timeline and State, in order
to re-trigger controllers when the timeline changes.
We can't just get rid of that behavior but we can inject it
from globals rather than having a direct coupling on globals.store.
That makes next CLs easier.
This allows to move TimelineManager from common/ to core/
next to the other managers.


Change-Id: I5a0a603cd1f9e3c1796c36c64c7f822d7ce5e2fc