test: Revamp integrationtest pre-main initialization
There are two cases when we want to run code before gtest starts running
all the tests:

1. heapprofd integrationtests
2. tracing api integrationtests

For 1, we used __attribute__(constructor), but this doesn't work well
anymore on android, probably because bionic is not fully initialized
yet (static order initialization is still a problem).

For 2, we put code directly in main, which was not ideal from an
isolation perspective.

This commit implements a simple mechanism to execute code in main before
gtest starts running.

Constraints:

* We want to run a minimal amount of code before main. Specifically, no
  allocation, STL. Just a global variable assignment.
* We want 1 and 2 to fail linking if they're used with a generic gtest
  main that doesn't have our mechanism.
* We don't want our main to depend directly on 1 and 2: in some build
  configurations 1 and 2 might be disabled.

Tested: run locally on standalone build
Tested: atest `perfetto_integrationtests` on a local cuttlefish

Bug: 363494560
Change-Id: I6f536960a55ca3a00ca049502eea70314f309d3d