Add support for perf aux data
This patch adds all the infrastructure necessary to parse aux data
contained in perf files. Aux data is just a stream of bytes that need to
be forwarded to a proper tokenizer that understands the data format.

What type of data is carried in the aux stream is determined by looking
at the AuxtraceInfo record. In this patch we have added a minimal
implementation for ETM data that just parses the configuration inside
the AuxtraceInfo record and discards all other data.

In the future we might add full support for ETM or support for other
types of data.

Change-Id: I612dd8fb21f69ac0478844315a11c25b5b419fe0