ui: add unified scrollTo
Scrolling to tracks/slices is very common all over the codebase.
Currently we have a plethora of functions with different entrypoints.
Furthermore "selection" and "scrolling to things" are quite intertwined.
This CL introduces a unified scrollTo function to be used everywhere.
This is NOT the final solution, as it seems in 90% of cases we want to
scrollTo something as a result of a selection, so perhaps it should be
selctionManager to do that, rather than callsites. But that is too big
of a refactoring, so i'm doing things in steps.
This helps cleaning up the dependency between plugins and frontend, and
brings some sanity unifying all the scrollTo variants into a single
entry-point.

Change-Id: I79436fb811e50d26ff64e2accc2538af433df432