shared_lib: Fix PerfettoDsImplReleaseInstanceLocked()
If PerfettoDsImplReleaseInstanceLocked() is executed while the data
source is being stopped, it can crash on a PERFETTO_CHECK.

Specifically the following series of events:
* The data source instance is active.
* [thread 1] The producer enters a tracepoint.
* [thread TracingMuxer] The producer receives a request to stop the data
  source: TracingMuxerImpl::StopDataSource_AsyncBeginImpl() is executed.
* [thread 1] PerfettoDsImplGetInstanceLocked() is called and succeeds,
  because DataSourceStaticState::valid_instances still has not been
  updated.
* [thread TracingMuxer] TracingMuxerImpl::StopDataSource_AsyncEnd() is
  executed. It will not be able to complete, because it requires the
  lock, but it will reset DataSourceStaticState::valid_instances.
* [thread 1] PerfettoDsImplReleaseInstanceLocked() is called.
  DataSourceStaticState::TryGet() will return nullptr because
  DataSourceStaticState::valid_instances has been reset.
* [thread 1] PERFETTO_CHECK is hit.

This commit fixes the problem by avoiding the TryGet(): since we're
holding the lock, we're sure that the instance is not destroyed, so we
can access that directly.

This is the simplest fix in terms of the minimal amount of changes to
the SDK internals. Other possible fixes would be:
* Resetting valid_instances under the lock
* Resetting valid_instances in StopDataSource_AsyncBeginImpl (but that
  prevents tracing from the OnStop callback).
But they're all more invasive in terms of changes to the C and C++ SDK.

This change only fixes the bug (which is only in the C SDK).

The included test reproduces by making some assumptions on the internal
implementation details of the SDK.

Bug: 358888843

Change-Id: Ib199fbef52390282ed4f992f489f26e7062eaf29