ui: Display CPU freq tracks even when sched events are not enabled
Currently the cpu_freq plugin gets the list of CPUs from the sched
table. So if ftrace sched events are not enabled, the CPU frequency
tracks will not be displayed. Get the list of CPUs from the
cpu_counter_track table instead.

Bug: 365912321
Test: Record traces both with and without ftrace sched events enabled
and check that the CPU freq tracks are displayed in both cases.

Change-Id: I1324357ad56117f851b82562013adca8f10bcee0