tp: switch tokenizer to real SQLite tokenizer + grammar
This CL prepares the ground for having a proper LALR parser by switching
to using the real sqlite tokenize function (instead of some franken
version of them) and also generating the token definitions directly
from SQLite instead of hacking it as "generic keyword".

All of this stuff will be used in the upcoming parser.

Change-Id: I0c62d9c3700f2e585609400e5d4b7742aed7825c