tp: fix initialization logic when someone else has initialized SQLite
This CL fixes the fact that trace processor is being a "bad citizen" and
crashing the embeddding process when we fail to initialize SQlite in the
way we want. Instead, gracefully respect the choice made.

Unfortunately the way to do this is very hacky and comes down to the
fact that we do not have a way to detect if someone else has already
initialized SQLite (i.e. there is no sqlite3_is_initialized). Because of
this, try to see if we can configure SQLite and if this failse, we can
safely assume it's because someone else already did it before us.

Change-Id: I2c664f949c1468cf92c2f503f2bbdc22223e7d34