traced: Capture CPU identifier
traced record many information which connected to CPUs in the system (PMUs,
thread, etc).

Unfortunately, no information to identify the CPU itself is recorded in the
trace. This prevent CPU-dependent processing to be applied in the trace
processor.

This change address this issue by recording an identifier when the
`linux.system_info` is enabled. The format of the identifier is CPU specific.
For Arm processors it follows the format
    "arm:<architecture>:<implementer>:<part number>:<variant>:<revision>"
Where architecture, implementer, part number, variant and revision are numbers
encoded in hexadecimal.

Change-Id: I098fb40582e865ee74f76a9332aa30329dc3cf91
Signed-off-by: Vincent Coubard <vincent.coubard@arm.com>