ui: Add discriminating type to {,Legacy}DetailsPanel
It's becoming a bit too confusing to handle the duality of
DetailsPanel vs LegacyDetailsPanel. In order to move everything
onto the new Trace API, it's more convenient if they have a
discriminating property, so we can have a single registerDetailsPanel()
function that can deal with both. This reduces churn in next CLs.

Change-Id: I483e23b095d1f9bd1e2e2706a3d916d5f8c2b960