tracing_service_impl: Emit lifecycle event with slow flush data sources
Some data sources can take a long time replying to a flush request
(on Android, possibly because their producer is frozen).

This commits records which (up to a limit) data sources haven't replied
to a flush request on timeout.

The data is recorded only for "final" flushes. periodic flushes are
excluded.

Change-Id: Ifdcc8c3b006667890593ab09c2fa22c61f04444c