ui: Add selection resolver API
If we are going to make the selection objects 'fat' i.e. contain
various pieces of information related to the entity that has been
selected, it's important to reduce the number of places where this
selection object is directly set, to reduce duplication and coupling.

This CL introduces a new selection resolver system where plugins can
register their ability to resolve selection objects on behalf of of
other parts of the UI. This is currently limited to SQL table names
and ids like so:

```
registerResolver('slice', (id) => {...selection object....});

// ~~~ snip ~~~

resolveSelection('slice', 123);
```

This could be made more generic way in the future, possibly using URIs,
but for now, sql table and ID are the only way folks look up selections
on other plugins, so this will do for now.

Change-Id: I297247629a7a504206b28ccb0a8ecdc72c7cff0c