ui: use the interrupt context when showing scheduling wakeups
> Per-cpu scheduling tracks at the top:
Before we would always render a wakeup link for all wakeup events,
which:
* did not distinguish between interrupt/non-interrupt wakeups, and
  would always blame the thread that was scheduled at the time.
* was actively wrong if the (interrupt) wakeup came from swapper at the
  time multiple cpus were idle, as the backing tables returned a random
  cpu index.

With these changes:
* for traces with irq_context, we draw only non-interrupt wakeups.
* for old traces, we draw non-swapper wakeups, attributing the wakeup to
  the scheduled thread.

If we wanted to keep visualising the interrupt wakeups, we'd need to add
waker_cpu to the thread_state table, at least to disambiguate the
swapper case.

> Per-thread state tracks:
Similarly, runnable slices' wakers and running slices' wakees weren't
using the interrupt context even if it was available.

With these changes:
* for traces with irq_context, non-interrupt wakeups are rendered as
  before. Interrupt wakeups instead are labeled as "Woken by Interrupt",
  without any extra visualisation or links.
* for old traces, we link to the thread state active when the wakeup
  happens, and add a "(maybe interrupt)" qualifier to the description.
* minor: the selected thread_state slice's details pane no longer duplicates
  the waker information.

Overall, this fixes wrong information being shown in the UI, but there are
still idiosyncrasies between cpu and thread tracks. I purposefully limited the
scope of changes since imo the right fix is to unify all of the wakeup logic
and reuse the same widgets, instead of having two parallel implementations.
Which would be a significant patch by itself.

Bug: 321039933
Change-Id: I9686d73eb72125dbca56a880e80ef1f6aae5bf84