ui: Only render timespan note durations when selected
Currently, timespan note durations are only rendered when the note is
NOT selected. While this made sense at the time for reasons, it makes
little sense now.

This CL inverts the logic, making notes only render their durations
when the ARE selected. Compared to rendering them all the time for all
notes, this helps stop the durations from overlapping one another and
becoming unreadable when lots of notes are present on screen.

Change-Id: Ia478fb1853c361e3ace08d23aeb96f9de317fac0