ui: Turn details panel hook on TrackDescriptor into a factory
Instead of the details panels being perpetual objects that are
constructed when the track descriptor is loaded, they are instead
supplied as factories, which are called each time the selection is
changed.

This simplifies the data loading story on each DetailsPanel
implementation as plugin authors no longer need to juggle old vs new
cached data while writing details panels, as well as removing the need
to manually do async limiting & schedule redraws when loading is
complete.

Whatsmore, details panels can now optionally supply an async load()
method, which is called (if defined) by the core just after creation,
and any loading should be done in here. This is used by the core to do
async limiting (avoids enqueueing too tasks if selection changes
quickly) and also for anti-flicker when loading is fast.

The reason for using an object with an async load() method rather than
just making the factory async, is because details panels mostly show
custom loading logic, and so we need a way to call render while the load is ongoing. If we used an async factory, the render() function wouldn't
be available until after the factory had completed. This might change if we switch to using a stock loading screen from the core instead
while details panels are loading.

Change-Id: I1993d08384300a75d60b9c9d59690765ff9a379b