tp: Add dimensions column to track_table
This is a big change but should be a noop, as the dimensions and classification columns are still hidden. The only visible difference now is adding things to args table, hence some changes to the tests.

The next step is to try and migrate all of the classification=kUnknown tracks and also remove python tables of all small child tracks tables and replace them with prelude SQL tables to maintain the API.

Change-Id: I0f604a97f8514eadda6a3511f31fb1b09a247567