stdlib: Remove redundant Wattson module
This is an old Wattson module that was made redundant by new metrics
added. All the functionality in this module is now encapsulated in
Wattson metrics. Additionally, the tables that are removed were too
rigid and didn't have the flexibility that is needed for external teams
to use.

This cleanup is an intermediate step to give the Wattosn modules a more
intuitive and understandable name.

Bug: 370829192
Test: tools/diff_test_trace_processor.py out/linux/trace_processor_shell --name-filter '.*wattson.*'
Change-Id: I467c1ff5cbc3a6a36be25a8ca1feafd03ab95f0e
Signed-off-by: Samuel Wu <wusamuel@google.com>