ui: Fix omnibox state restoration bug
When setting the omnibox status message, the previous omnibox state is
remembered and restored after a timeout. However, this ignores the fact
that the state might be changed in the meantime, making it possible for
the state to be restored back to the wrong thing.

This CL patches the bug by creating a separate orthogonal state for the
presence of the status message. This means we never have to save and
restore any state, which means we can never get it wrong.

Bug: 372650586
Change-Id: I9aef8628dc298b1235f76524709caee946458f96