ui: Remove remaining unnecessary scrollTos at selectSqlEvent callsites
Before selectSqlEvent(), it was customary to make selections then
manually scroll to the selection by finding the track URI and invoking
scrollTo().

Since selectSqlEvent(), this can be achieved by passing the
`scrollToSelection` flag to instead.

This CL does just this at all the remaining selectSqlEvent callsites,
which removes a lot of redundant code.

Change-Id: I84a26ccad44ee2f1e58e32b3481671462124cc91