ui: Remove legacy selection types from serialization code
Note: This CL breaks compatibility with permalinks from older versions
of the UI which contain one of these legacy selection types.

This is not a disruptive as it used to be since Primiano's explicit
deserialization change - i.e. the trace can still be loaded but none of
the permalink data will be restored.

Change-Id: Ibf96bf84056a7766acbc389c62d6b034640989a0