ui: Add area selections to permalinks
Note: This CL breaks compatibility with older versions of the UI when
permalinks are created containing area selections.

This is not as disruptive as it used to be since Primiano's explicit
deserialization change - i.e. the trace can still be loaded but none of
the permalink data will be restored.

Change-Id: I880eb3a8c052fefc5d394e916dcad817ff9c1eb4