ui: tp: remove legacy derived events from trace processor
This CL deletes all remaining derived events from trace processor and
the UI as well as all supporting code. Most of the important events
were already migrated to plugins, remaining events have not been touched
in years so likely are not meaningful anymore and it's not worth us
supporting them further.

Change-Id: Iccced0c0789422403e24017410ad71ddf08496af