tracing_service: Implement clone by unique session name
Today, the only way to clone an existing tracing session is by using its
id.

The id of a tracing session can be found by either:
* Listing all the active tracing sessions with QueryServiceState.
* Receiving an Observable event for a clone trigger.

Other than these a consumer has no way of knowing its own active tracing
session id.

This offers a way of using the unique_session_name passed in the config
to choose the tracing session to clone.

WARNING: this is a breaking change for embedders who depend on `ext/`
(chromium).

Change-Id: Ic43d183768dffe59dac355aa3b70223838c45677