ui: remove JobStatusUpdate and simplify sidebar logic
This CL simplifies the way deferred conversion jobs
are handled AND the sidebar code.
The end user feature is the following: some of the actions
accessible by the sidebar (convert to json or legacy ui)
take several seconds and are offloaded to a worker. Today
we show a spinner on the sidebar when these are in progress.
Today the way the progress tracking works involves a lot of
plumbing all over the places (globals, publish, etc).
Instead I'm changing this as follows:
- I rewrote the long conversion operations as promises, which
  resolve when the conversion is done (which is notified by
  a postmessage from the dedicated worker)
- I updated the spinner logic in the sidebar to apply to any
  case when the action of the menu item returns a promise.
  This removes a lot of special-case handling for those spinners.

Change-Id: I41fc99be10cb789d230551220da66b0f00db7531