ui: fix SearchOverviewTrack race
Problem: createSearchOverviewTrack() is an async method and we
used to put it in the map only after the promise resolved. This
meant that if we switched pages before that, we would interleave
a second call to it, which would clash with the original.
As part of fixing this, turn the SearchOverviewTrack into an
actual class. The existing createSearchOverviewTrack() function
was a function that contains nested functions, contains variables
that get bound and accessed by different sub-functions, and returns
an interface with methods. That is by all means a class. Given this
is not 2015 anymore, let's use a class construct for it.

Bug: 375342066
Change-Id: Ic7709c2bebfcc0426773010d5d78fdaf870d5948