ui: Fix track error capturing logic
Currently, if a track crashes in one of its lifecycle hooks, the
hook will more often than not be called once more.

This is particularly a problem when a track crashes in its onCreate
function, it more than likely leaves the track in a partially
initialized state where some views and tables have been created.

If we call onCreate again, we usually end up trying to recreate these
same tables and end up crashing again but for a different reason.

Because we only store the most recent error, this means that only the
duplicate table crash is reported to the user, hiding the reason for the
original crash.

This CL patches the issue by:
- Returning early from the scheduled task if an error is present
- Catching the error using a try-catch block in the scheduled task,
  rather than using a .catch() handler on the promise, which ensures
  the error is set before the next task runs, rather than being
  deferred to the microtasks.

Change-Id: I9f5066a0e2689ee639a3cd852c89c355e2ee7613