ui: gate child area selection based on summary track
We have some logic with area selection to make area selections on
summary tracks automatically select an area in all children.

However, this logic was gated on whether it was a child track *not*
whether it is a summary track. This leads to confusing behaviour for
tracks which actually have events on them.

Change-Id: I0f2b8aa5012713a4d5b6b13c8e7672762df7e5e1