Increase kprobes maxactives
To handle the case where the pair kprobe/kretprobe is attached to a
functions that can be executed concurrently, for example because called
at the same time on different CPUs or because the calling process can be
preempted and a different process can call the same function, the kernel
needs to keep a pool of instances to track the execution stages of every
kprobe. The pool size is defined, per kretprobe, by the maxactive
parameter. If the maxactive number is too small, the pool may overflow
and kretprobe events may not find their matching kprobes, being missed.

To solve this issue, increase the default number of maxactives from the
default value (from the kernel documentation, if CONFIG_PREEMPT is
enabled, the default is max(10, 2*NR_CPUS)) to 1024. This number has
been found empirically, to reduce the number of misses for the Android
functions binder_ioctl_write_read and binder_ioctl, identified as being
problematic for their preemptability and popularity in Android.

Test: Visual inspection of /sys/kernel/tracing/kprobe_profile
Bug: 359542775
Change-Id: Ic74b0d3a5374c53be75896d4e9355dd65f787a46
Signed-off-by: Alessio Balsini <balsini@google.com>