Export kprobe_profile stats
Whenever kretprobes cannot match a previously generated kprobe event
because of limited pool size set with maxactive parameter, a counter of
missed events is incremented, and is made available to userspace through
/sys/kernel/tracing/kprobe_profile.
This patch parses the kprobe_profile file and exports cumulative values
for all the hits and misses, exported as part of FtraceStats.  If the
delta of misses is greater than zero, the trace will be flagged as
having a data loss.

Test: perfetto_unittests
Bug: 359542775
Change-Id: I1bef35f54068969087fa9032401a0638dd6c85bc
Signed-off-by: Alessio Balsini <balsini@google.com>