ui: remove last dependency between globals > Trace
This removes the last accessor that was using globals.trace,
allowing to remove the _initialFakeTrace hack.
Doing so requires some type casting in some frontend classes.
This is the problem:
- Various plugins (that see only Trace and not TraceImpl)
  use NamedSliceTrack / ThreadSliceTrack.
- So plugins can only pass Trace into them.
- Internally, the current implementation of
  {Name,Thread}SliceTrack uses ThreadSliceDetailsTab.
- ThreadSliceDetailsTab needs to access flows.
- Flows are only accessible via TraceImpl and not Trace.
So we have two options here:
1. Casting Trace->TraceImpl inside {Name,Thread}SliceTrack
2. Exposing flows on the public Trace interface

I did 1 because I feel we might want to rethink flows a bit
later on and I didn't feel exposing them.

Change-Id: I269fae57e7b6e0705ba3c0c1baaf66b5dd94f869