ui: allow any App to obtain the matching Trace for the plugin
Why am I doing this change: in order to properly support
pages-as-plugins I need a way to go from App(pluginFoo) to
Trace(pluginFoo).
Today the lifetime of TraceImpl is a bit odd: AppImpl has a

currentTrace: TraceImpl member, but that member is set only in
the core instance of AppImpl and is always undefined in the
per-plugin instance.
This CL moves the knowledge of "there are N AppImpl, one for each
plugin" into AppContext, and similarly to TraceImpl<>TraceContext.
Change-Id: Ie2f852909beb2b4898371b606652c81a704ef788