Deflake HeapprofdJavaCtsTest Runtime tests
The HeapprofdJavaCtsTest sometimes fail.

This happens because the app process, while forking, could hold locks in
other threads. Since other threads are not part of the fork, they will
not release locks in the forked process. This is a known limitation of
the feature that we're testing.

In order to avoid hitting this case in the test, we can reduce the work
that the test app is doing.

Let's introduce a new empty activity just for the HeapprofdJavaCtsTest.

Bug: 356328336
Change-Id: I827aa358b4a94b60edc2197fe1824ce68dbc8e15