tp: add table for breaking down thread_slice by scheduling states
This CL adds a new module `slices.time_in_state` which contains a table
breaking down the thread_slice table by each scheduling state. This
allows for simpler queries where you want to summarise the running
behaviour of a slice in simpler terms.

We also change the sched.time_in_state module to indicate that it's not
designed for large number of intervals (tbh this is something we should
really fix...) and to change some documentation to better match the
slice documentation.

Change-Id: I787a4b9592d860170ff164456b742235ac067111
Fixed: 2