Make power counter always positive
The current counter is negative when discharging, but it's more useful
for the power counter, and therefore the average power metric derived
from it, to always be positive. This is also consistent with other
sources which report positive power regardless of charging/discharging,
e.g. ODPM and Intel RAPL (not supported in perfetto yet).

Bug: 378710178
Test: ui/run-dev-server, check batt.power_mw counter is positive even
when current is negative.

Change-Id: Ic68e35ee83413a177ac22ade758164acff621907