wattson: Fix estimates when entire CPU track missing
When an entire CPU freq or idle track is missing, then some other CPU
estimates won't be calculated.

Add a NULL but defined time range for CPUs that are physically present
on the device but don't have a single freq/idle event triggered. This is
so that interval_intersect() operates on a valid time range.

Test: tools/diff_test_trace_processor.py out/linux/trace_processor_shell --name-filter '.*wattson.*'
Bug: 379160886
Change-Id: I3b3fa24a61c2e348802b064e17d7dc93000f32b2
Signed-off-by: Samuel Wu <wusamuel@google.com>