Add new table capturing layer_id associated with multiple actual frames.
This will become the base table, on which the android_frames table will
be based. Additionally, this restructuring allows for the correction of
inaccurate assumptions in the old `android_frames` table. This involves
the removal of certain fields and improvements to the documentation of
others. More information on the tables captured here:
go/android-frames-perfetto-changes

Test: tools/diff_test_trace_processor.py --keep-input --name-filter "Frames:android_layer_frames" out/linux_clang_debug/trace_processor_shell
Change-Id: I4e0cf45ba271e7d5cc9d8c0fdaa7634717f700cd
Signed-off-by: Devarshi Bhatt <devarshimb@google.com>