ui: Use plugin deps to add tracks to process groups
Currently, plugins that add tracks to process or track groups get
access to the group by looking up these groups via their well known ids.
The problem is getTrackById isn't particularly efficient and is called
frequently during trace load.

This CL switches to using plugin dependencies to solve the issue of
multiple plugins adding to a shared group instead.

We have the ProcessThreadGroups plugin which exposes a
getGroupForProcess and getGroupForThread method. Other plugins can use
these methods to look up the correct group for their track, and the
dependencies list means that that this plugin is run before any
dependent plugin so these groups are guraranteed to exist by the time
the dependent plugins are run.

Change-Id: Ic1c3e077c2fbe7f145401c361e5e408a4da7a908