ui: Load Wattson on track expansion
Only start loading Wattson when user clicks on Wattson track. The motive
is so that normal UI loading of other Perfetto operations can happen on
their own timeline, and for the more expensive Wattson calculations only
to happen if a user wants Wattson information. This is a prerequisite
step to enable Wattson plugin by default.

Bug: 364902925
Change-Id: I5bfb8642a3d6b3c26cc0741fa14fcd78161f4d7a
Signed-off-by: Samuel Wu <wusamuel@google.com>