tp: Parse Battery Stats History items from checkin
Parse the Battery Stats history items from the checkin included in bug
reports. This CL attempts to emit similar tracks, events, and counters
to what the Atrace implementation in BatteryStatsHistory.java will emit:
https://cs.android.com/android/platform/superproject/main/+/main:frameworks/base/core/java/com/android/internal/os/BatteryStatsHistory.java;l=1762-1770;drc=20c4b0d30c67e4acf89d87ac7b5e1d8b83365fc0

Notable changes from the atrace implemenation include:
- all the battery events are not on global tracks. The checkin does not
 provide PIDs, so there is no PID with which to create a process track.
- the "battery_stats.running" is included.

This CL does not attempt to parse or emit events that are in the battery
stats checking, but not already emitted by that Atrace code. This
includes, but is not limited to:
Wakelocks, package/uid info, and summary stats.

Inclusion of this data will be left to a later CL.

Change-Id: Ie92578df34041595c0caaa0323eb4fe91e2784aa