ui: Speed up classNames() by 10x
classNames is a purvasive function that is called fairly frequently
during render cycles, espeically when we have lots of tracks.

While it doesn't take up a significant fraction of the render time, it
is near the top of the list when sorting by self-time in devtools.

This CL optimizes this tiny function to within an inch of its life
(benchmarked on chrome) so that now it barely registers in the profiler
at all.

It also changes the functionality slightly as it removes support for
flattening args as no one was really depending on this anyway.

Change-Id: I328f99aad7fccf9cc47b180b80f7e37cc018d6a0