ui: Resolve track_id -> TrackDescriptor on selection change
Currently we build a map of track_ids -> TrackNodes every canvas
render cycle, which can take a long time when there are a lot of tracks
on the workspace.

This CL improves on this by resolving track_ids to URIs just once
when flow events are loaded (i.e. when the selection changes), which
means we now only need to build a map of track URIs -> TrackNodes
every render cycle which is significantly more efficient.

It's safe to resolve track URIs at selection change time as we can
assume that the track registry is relatively static after trace load,
and we also make this assumption elsewhere as we load track information
for TrackEvent selections on selection change too.

Change-Id: Ie2f331618b68d66d44350aa10335cad58435bafd