ui: Allow plugins to define scss files
This CL adds the ability for plugins to define styles in their own dirs
without having to touch the core assets directory. This helps to keep
the entire plugin definition in one place. The need for this is
especially pressing since plugins can now define pages which usually
necessitate custom styles.

Plugins can use this feature by adding a `styles.scss` file at the root
of their directory. This file will be included into the bundled css
when the UI is built.

Note: If running `ui/run-dev-server`, it will need to be restarted when
adding or removing a styles.scss, or when adding a new plugin.

Change-Id: Ic07e35cd308c93a56cf62f0b090b02018e49357c