[perfetto] Fix path parsing in tools/check_sql_modules.py
Before the CL:
For each input file with an error, the script tried to print its path
relative to ${PERFETTO_DIRECTORY}/tools [1][2].

Problem:
The input files don't always live under ${PERFETTO_DIRECTORY}/tools,
see the --stdlib-sources flag [3]. This led to IndexError in [1].

After the CL:
The error message displays the full path of the problematic files. I
I considered using the path relative to --stdlib-sources as an
alternative, but that'd be confusing when the flag is omitted and
its default value is used.

[1] https://source.chromium.org/chromium/chromium/src/+/main:third_party/perfetto/tools/check_sql_modules.py;l=135;drc=8d4918c9b60d1b7865499911f2cb0bdb5151cd3f
[2] https://source.chromium.org/chromium/chromium/src/+/main:third_party/perfetto/tools/check_sql_modules.py;l=25;drc=8d4918c9b60d1b7865499911f2cb0bdb5151cd3f
[3] https://source.chromium.org/chromium/chromium/src/+/main:third_party/perfetto/tools/check_sql_modules.py;l=39-42;drc=8d4918c9b60d1b7865499911f2cb0bdb5151cd3f

Bug: 380836769
Change-Id: I57d1b8ab99bb5943992d709e24a1de071ff6c94f