profiling: Fix an unwinding fail issue on RISC-V64
Due to the difference in calling conventions, a frame address cannot be used
for getting a correct stack pointer value. See following assembly codes:

  RISC-V64:
    0x3f8639ad62:    add    sp,sp,-544
    0x3f8639ad66:    sd    ra,536(sp)
    0x3f8639ad6a:    sd    s0,528(sp)
    0x3f8639ad6e:    add    s0,sp,544

  AARCH64:
    0x7f79bd36b0:    sub    sp, sp, #0x210
    0x7f79bd36b4:    stp    x29, x30, [sp]
    0x7f79bd36b8:    mov    x29, sp

I've got these assembly codes by using GCC-14 (for RISC-V) and GCC-9 (for AARCH64).
Both assembly codes are first few lines of RecordMalloc for each denoted
architecture. In RISC-V, an s0 (used for storing the frame pointer) register has
a caller's stack pointer value while an x29 (also used for the frame pointer)
register in AARCH64 has a modified stack pointer value.

Because of that, stack-unwinding in RISC-V is failed when it tries to get
the first return address (which is stored in sp + 8).

This patch changes the method to get the stack pointer to use the value stored
in the sp register.

This patch is related to the following github issue report:

  https://github.com/google/perfetto/issues/936

Signed-off-by: Sung-hun Kim <sfoon.kim@samsung.com>
Change-Id: Ia7c677fb70eec312e57d7f5fa565273bafabb6a5