Close broken suspend_enter slice earlier.
Add a little state machine to detect b/381038940 and close the slice when we do so, rather than waiting for the next suspend event.

Remove the more general auto-closer based on `ongoing_suspend_resume_actions`. This seems to be the only real broken case.

Remove `suspend_resume_cookie_map_` as well; as far as I can determine we don't need it; we can just generate cookies from StringIds when we need them.

Bug: 381039361
Change-Id: I778d7d5bc9c2a00bebb43ab4512103da81b8b198