shared_memory_abi: Rename PageHeader::layout to header_bitmap
PageHeader::layout is a 32-bit bitmap that includes the state of each
chunk and the layout of the page.

The fact that the term `layout` is used both for the whole bitmap and
for the 3 bits that encode the page layout is kind of confusing.

This commit renames `layout` to `header_bitmap`. The remaining uses of
layout only refer to the 3-bit page layout.

This commit also makes use of helper functions where possible, instead
of doing bit shifts directly.

Change-Id: Icac201d76a8ba441a857f583ed8ac49585fb3533