ftrace: bump "have we caught up to the writer" heuristic to half a page (from ~3.5k bytes)
Since non-blocking reads from ftrace do not let us detect how much data
is left in the buffer, we use a heuristic that decides that we've caught
up to the writer if we see a under-filled tracing page (the data is
always framed as 4k tracing pages, but the header says how much of it is
actual events). This can have false positives if there are large events
that cannot fit until the start of the next page. We've used "page minus
512 bytes" so far, but unfortunately android userspace tracing via
ftrace/print is known to be logging strings up to 1kb in size at the top
end.

Normally false positives aren't a deal breaker since a future read pass
collects the subsequent pages, but they can be a bigger problem for the
final flush, since that's where the trace stops. This recently cropped
up in a case where ftrace was being configured with very long read
periods (>10s), putting way more importance on the final flush to read a
large volume of data. This was a brittle configuration, but we can still
try to decrease the false positive rate as it also helps prevent
truncated ReadTicks.

After evaluating several options, we've decided to simply change the
heuristic to "half of a page", lowering the chance of false positives to
cases with >2kb event sizes, which are highly unlikely on most devices.

Alternatives considered, but rejected (for now):
* using tracing buffer poll() - while the eventual goal, we cannot yet
  assume that most devices have kernels with recent fixes to this
  feature.
* parsing "number of bytes in buffer" from tracefs/per_cpu/cpuX/stats, as
  it's known to be inaccurate until at least least v6.6
  (https://github.com/torvalds/linux/commit/45d99ea). This generally
  makes me less keen on depending on buffer stats for core
  functionality.
* parsing "number of events in buffer" stats. Our code isn't really
  structured for counting events (as that is only known once the page
  payloads are parsed), and the associated refactoring isn't worth the
  effort for this issue.
* looking for multiple consecutive pages that are under-filled. While
  it'd be an improvement, we're not expecting events above 2kb in the
  first place, so this feels like an unnecessary addition.

Bug: 378651171
Change-Id: I80e2bc6bcf21a7b097ace90fdaccc45f68e4e5bf