ui: fix recent wakeup info regression in thread_state and sched tabs
Addressing two issues with r.android.com/3344159:
1) typo where waker_id was being used as waker_utid, causing the "sched"
wakeup info to be bizzare (empty or a thread whose id coincided with the
thread_state id of the sched slice).
2) when looking up waker thread_state for a running thread_state slice,
the code was still using the running slice's wakerInterruptCtx, instead
of the previous runnable's. This was rendered as if the trace predates
interrupt context, stating that we're unsure whether the wakeup is
indeed from that thread, or from an overlapping irq.

Fixes: 3c499f398f ui: Render additional thread state information
Bug: 199431666
Change-Id: Ib2907585b129e89913a23c814810f5c2a9389c52