tp: Populate ScreenState track from dumpstate battery_stats
The ScreenState track is a special track that is recognized by existing
plugins and is currently populated from 2 data sources:
- android.power data source (for inital screen state)
- "power" atrace category - for screen state changes.

This CL uses the screen change events in the battery stats checkin
within dumpstate to populate the same ScreenState track. This CL does
not attempt to populate the initial ScreenState value and only emits
screen changes, as initial screen state info does not appear to
be present in battery stats and may require other AOSP changes to get
that info.

Change-Id: Icc2adf3d4612a9ddf4890faf59469db9c9599c16