stdlib: Disable JOINID validation
Current solution require whole stdlib to be available, which is not true both for Chrometto and google3 modules. Because of this, having this check in any place that doesn't need whole standard library is a bad approach and we should move it to be runtime in Trace Processor.

Next step is to save all ID columns in perfetto_sql_parser/engine and then verify all joinid columns there with similar check to the one removed here.

Bug:381538049
Change-Id: Ie962f8358a8e0844b91ad5f274d4ab6c727fc465