stdlib: Add singleton wakeup power attribution
For singleton threads, threads that wake up a CPU to process only one
thread, define the idle transition cost as the sum of the idle exit cost
and the idle entry cost. This means adding the swapper cost before and
after the single thread wakeup, for the idle exit and idle entry power
usage respectively.

Test: tools/diff_test_trace_processor.py out/linux/trace_processor_shell --name-filter '.*wattson.*
Bug: 382070419
Change-Id: I0984524f8985297e239c2d6a355da24ae9765039