weak_runner.h: Move from src/ to include/ext/
I recently introduced `weak_runner.h` in `src/base/` and used it in
`src/tracing`.

Apparently bazel, in some build configurations with layering_check is
not happy that a header listed in the `sources` of `src_base_base` is
included from src_tracing.

Fix the problem by putting the header in include/perfetto/ext instead.

The downside is that headers in include/perfetto/ext can contractually
be included from chromium and android as well.

Change-Id: I79fb2289218a68bfc32bbf71dde3fa1a875cd815