base: add StringViewSplitter utility
Add a StringViewSplitter utility class similar to the existing
StringSplitter class, except that it can operate on StringView
objects without any copying of the string contents.

The existing StringSplitter can't be used for this purpose since in
order to use the existing StringSplitter with a StringView, it
requires converting to std::string first, which create a new copy.

This new class is meant to closely match the API and behavior of the
existing StringSplitter with a few differences:
- it operates on StringView objects
- cur_token() returns and empty string in cases where the original
 would return nullptr
- cur_tok_size() was removed since the caller can now simply do
 cur_token().size()
- Existing unit tests were ported and reused, but with changes that
 allow it to confirm with the new StringViewSplitter API

Change-Id: I8a29cd084b1947f530f7ae3539371f12815051a8